Abstracted out WA stuff into separate classes
