+++ WhatsAppApi/WhatsApp.cs
@@ -20,220 +20,13 @@ namespace WhatsAppApi
     /// <summary>
     /// Main api interface
     /// </summary>
-    public class WhatsApp
+    public class WhatsApp : WhatsAppBase
     {
-
-        /// <summary>
-        /// Describes the connection status with the whatsapp server
-        /// </summary>
-        public enum CONNECTION_STATUS
-        {
-            UNAUTHORIZED,
-            DISCONNECTED,
-            CONNECTED,
-            LOGGEDIN
-        }
-
-        private ProtocolTreeNode uploadResponse;
-    
-        /// <summary>
-        /// An instance of the AccountInfo class
-        /// </summary>
-        private AccountInfo accountinfo;
-
-        /// <summary>
-        /// Determines wether debug mode is on or offf
-        /// </summary>
-        public static bool DEBUG;
-
-        /// <summary>
-        /// The imei/mac adress
-        /// </summary>
-        private string imei;
-
-        /// <summary>
-        /// Hide online status
-        /// </summary>
-        protected bool hidden;
-
-        /// <summary>
-        /// Holds the login status
-        /// </summary>
-        private CONNECTION_STATUS loginStatus;
-
-        public CONNECTION_STATUS ConnectionStatus
-        {
-            get
-            {
-                return this.loginStatus;
-            }
-        }
-
-        protected KeyStream outputKey;
-
-        /// <summary>
-        /// A lock for a message
-        /// </summary>
-        private object messageLock = new object();
-
-        /// <summary>
-        /// Que for recieved messages
-        /// </summary>
-        private List<ProtocolTreeNode> messageQueue;
-
-        /// <summary>
-        /// The name of the user
-        /// </summary>
-        private string name;
-
-        /// <summary>
-        /// The phonenumber
-        /// </summary>
-        private string phoneNumber;
-
-        /// <summary>
-        /// An instance of the BinaryTreeNodeReader class
-        /// </summary>
-        private BinTreeNodeReader reader;
-
-        /// <summary>
-        /// The timeout for the connection with the Whatsapp servers
-        /// </summary>
-        private int timeout = 300000;
-
-        /// <summary>
-        /// An instance of the WhatsNetwork class
-        /// </summary>
-        private WhatsNetwork whatsNetwork;
-
-        /// <summary>
-        /// Holds the encoding we use, default is UTF8
-        /// </summary>
-        public static readonly Encoding SYSEncoding = Encoding.UTF8;
-
-        /// <summary>
-        /// Empty bytes to hold the challenge
-        /// </summary>
-        private byte[] _challengeBytes;
-
-        /// <summary>
-        /// Default class constructor
-        /// </summary>
-        /// <param name="phoneNum">The phone number</param>
-        /// <param name="imei">The imei / mac</param>
-        /// <param name="nick">User nickname</param>
-        /// <param name="debug">Debug on or off, false by default</param>
         public WhatsApp(string phoneNum, string imei, string nick, bool debug = false, bool hidden = false)
         {
-            this.messageQueue = new List<ProtocolTreeNode>();
-            this.phoneNumber = phoneNum;
-            this.imei = imei;
-            this.name = nick;
-            this.hidden = hidden;
-            WhatsApp.DEBUG = debug;
-            this.reader = new BinTreeNodeReader();
-            this.loginStatus = CONNECTION_STATUS.DISCONNECTED;
-            this.BinWriter = new BinTreeNodeWriter();
-            this.whatsNetwork = new WhatsNetwork(WhatsConstants.WhatsAppHost, WhatsConstants.WhatsPort, this.timeout);
+            this._constructBase(phoneNum, imei, nick, debug, hidden);
         }
 
-        /// <summary>
-        /// Add a message to the message que
-        /// </summary>
-        /// <param name="node">An instance of the ProtocolTreeNode class</param>
-        public void AddMessage(ProtocolTreeNode node)
-        {
-            lock (messageLock)
-            {
-                this.messageQueue.Add(node);
-            }
-
-        }
-
-        /// <summary>
-        /// Connect to the whatsapp network
-        /// </summary>
-        public void Connect()
-        {
-            try
-            {
-                this.whatsNetwork.Connect();
-                this.loginStatus = CONNECTION_STATUS.CONNECTED;
-                //success
-                if (this.OnConnectSuccess != null)
-                {
-                    this.OnConnectSuccess();
-                }
-            }
-            catch (Exception e)
-            {
-                if (this.OnConnectFailed != null)
-                {
-                    this.OnConnectFailed(e);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Disconnect from the whatsapp network
-        /// </summary>
-        public void Disconnect(Exception ex = null)
-        {
-            this.whatsNetwork.Disconenct();
-            this.loginStatus = CONNECTION_STATUS.DISCONNECTED;
-            if (this.OnDisconnect != null)
-            {
-                this.OnDisconnect(ex);
-            }
-        }
-
-        /// <summary>
-        /// Encrypt the password (hash)
-        /// </summary>
-        /// <returns></returns>
-        public byte[] encryptPassword()
-        {
-            return Convert.FromBase64String(this.imei);
-        }
-
-        /// <summary>
-        /// Get the account information
-        /// </summary>
-        /// <returns>An instance of the AccountInfo class</returns>
-        public AccountInfo GetAccountInfo()
-        {
-            return this.accountinfo;
-        }
-
-        /// <summary>
-        /// Retrieve all messages
-        /// </summary>
-        /// <returns>An array of instances of the ProtocolTreeNode class.</returns>
-        public ProtocolTreeNode[] GetAllMessages()
-        {
-            ProtocolTreeNode[] tmpReturn = null;
-            lock (messageLock)
-            {
-                tmpReturn = this.messageQueue.ToArray();
-                this.messageQueue.Clear();
-            }
-            return tmpReturn;
-        }
-
-        /// <summary>
-        /// Checks wether we have messages to retrieve
-        /// </summary>
-        /// <returns>true or false</returns>
-        public bool HasMessages()
-        {
-            if (this.messageQueue == null)
-                return false;
-            return this.messageQueue.Count > 0;
-        }
-
-        /// <summary>
-        /// Logs us in to the server
-        /// </summary>
         public void Login(byte[] nextChallenge = null)
         {
             //reset stuff
@@ -272,23 +65,6 @@ namespace WhatsAppApi
             this.SendAvailableForChat(this.name, this.hidden);
         }
 
-        protected void SendData(byte[] data)
-        {
-            try
-            {
-                this.whatsNetwork.SendData(data);
-            }
-            catch (ConnectionException)
-            {
-                this.Disconnect();
-            }
-        }
-
-        /// <summary>
-        /// Send a message to a person
-        /// </summary>
-        /// <param name="to">The phone number to send</param>
-        /// <param name="txt">The text that needs to be send</param>
         public void Message(string to, string txt)
         {
             var tmpMessage = new FMessage(GetJID(to), true) { data = txt };
@@ -322,38 +98,6 @@ namespace WhatsAppApi
             this.SendNode(node);
         }
 
-        /// <summary>
-        /// Convert the input string to a JID if necessary
-        /// </summary>
-        /// <param name="target">Phonenumber or JID</param>
-        public static string GetJID(string target)
-        {
-            if (!target.Contains('@'))
-            {
-                //check if group message
-                if (target.Contains('-'))
-                {
-                    //to group
-                    target += "@g.us";
-                }
-                else
-                {
-                    //to normal user
-                    target += "@s.whatsapp.net";
-                }
-            }
-            return target;
-        }
-
-        /// <summary>
-        /// Send an image to a person
-        /// </summary>
-        /// <param name="msgid">The id of the message</param>
-        /// <param name="to">the reciepient</param>
-        /// <param name="url">The url to the image</param>
-        /// <param name="file">Filename</param>
-        /// <param name="size">The size of the image in string format</param>
-        /// <param name="icon">Icon</param>
         public void MessageImage(string to, string filepath)
         {
             to = GetJID(to);
@@ -387,7 +131,7 @@ namespace WhatsAppApi
             }
 
             //request upload
-            UploadResponse response = this.UploadFile(filehash, "image", finfo.Length, filepath, to, type);
+            WaUploadResponse response = this.UploadFile(filehash, "image", finfo.Length, filepath, to, type);
 
             if (response != null && !String.IsNullOrEmpty(response.url))
             {
@@ -431,7 +175,7 @@ namespace WhatsAppApi
             }
 
             //request upload
-            UploadResponse response = this.UploadFile(filehash, "video", finfo.Length, filepath, to, type);
+            WaUploadResponse response = this.UploadFile(filehash, "video", finfo.Length, filepath, to, type);
 
             if (response != null && !String.IsNullOrEmpty(response.url))
             {
@@ -483,7 +227,7 @@ namespace WhatsAppApi
             }
 
             //request upload
-            UploadResponse response = this.UploadFile(filehash, "audio", finfo.Length, filepath, to, type);
+            WaUploadResponse response = this.UploadFile(filehash, "audio", finfo.Length, filepath, to, type);
 
             if (response != null && !String.IsNullOrEmpty(response.url))
             {
@@ -493,46 +237,7 @@ namespace WhatsAppApi
             }
         }
 
-        private byte[] CreateThumbnail(string path)
-        {
-            if (File.Exists(path))
-            {
-                Image orig = Image.FromFile(path);
-                if (orig != null)
-                {
-                    int newHeight = 0;
-                    int newWidth = 0;
-                    float imgWidth = float.Parse(orig.Width.ToString());
-                    float imgHeight = float.Parse(orig.Height.ToString());
-                    if (orig.Width > orig.Height)
-                    {
-                        newHeight = (int)((imgHeight / imgWidth) * 100);
-                        newWidth = 100;
-                    }
-                    else
-                    {
-                        newWidth = (int)((imgWidth / imgHeight) * 100);
-                        newHeight = 100;
-                    }
-
-                    Bitmap newImage = new Bitmap(newWidth, newHeight);
-                    using(Graphics gr = Graphics.FromImage(newImage))
-                    {
-                        gr.SmoothingMode = System.Drawing.Drawing2D.SmoothingMode.HighQuality;
-                        gr.InterpolationMode = System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic;
-                        gr.PixelOffsetMode = System.Drawing.Drawing2D.PixelOffsetMode.HighQuality;
-                        gr.DrawImage(orig, new Rectangle(0, 0, newWidth, newHeight));
-                    }
-                    MemoryStream ms = new MemoryStream();
-                    newImage.Save(ms, System.Drawing.Imaging.ImageFormat.Jpeg);
-                    ms.Close();
-                    return ms.ToArray();
-                }
-            }
-            return null;
-        }
-
-        private UploadResponse UploadFile(string b64hash, string type, long size, string path, string to, string contenttype)
+        private WaUploadResponse UploadFile(string b64hash, string type, long size, string path, string to, string contenttype)
         {
             ProtocolTreeNode media = new ProtocolTreeNode("media", new KeyValue[] {
                 new KeyValue("hash", b64hash),
@@ -556,7 +261,7 @@ namespace WhatsAppApi
             }
             if (this.uploadResponse != null && this.uploadResponse.GetChild("duplicate") != null)
             {
-                UploadResponse res = new UploadResponse(this.uploadResponse);
+                WaUploadResponse res = new WaUploadResponse(this.uploadResponse);
                 this.uploadResponse = null;
                 return res;
             }
@@ -637,7 +342,7 @@ namespace WhatsAppApi
                         {
                             string fooo = line.TrimEnd(new char[] { (char)0 });
                             JavaScriptSerializer jss = new JavaScriptSerializer();
-                            UploadResponse resp = jss.Deserialize<UploadResponse>(fooo);
+                            WaUploadResponse resp = jss.Deserialize<WaUploadResponse>(fooo);
                             if (!String.IsNullOrEmpty(resp.url))
                             {
                                 return resp;
@@ -651,56 +356,6 @@ namespace WhatsAppApi
             return null;
         }
 
-        public class UploadResponse
-        {
-            public string url { get; set; }
-            public string mimetype { get; set; }
-            public int size { get; set; }
-            public string filehash { get; set; }
-            public string type { get; set; }
-            public int width { get; set; }
-            public int height { get; set; }
-
-            public int duration { get; set; } 
-            public string acodec { get; set; }
-            public int asampfreq { get; set; }
-            public string asampfmt { get; set; }
-            public int abitrate { get; set; }
-
-            public UploadResponse()
-            { }
-
-            public UploadResponse(ProtocolTreeNode node)
-            {
-                node = node.GetChild("duplicate");
-                if (node != null)
-                {
-                    int oSize, oWidth, oHeight, oDuration, oAsampfreq, oAbitrate;
-                    this.url = node.GetAttribute("url");
-                    this.mimetype = node.GetAttribute("mimetype");
-                    Int32.TryParse(node.GetAttribute("size"), out oSize);
-                    this.filehash = node.GetAttribute("filehash");
-                    this.type = node.GetAttribute("type");
-                    Int32.TryParse(node.GetAttribute("width"), out oWidth);
-                    Int32.TryParse(node.GetAttribute("height"), out oHeight);
-                    Int32.TryParse(node.GetAttribute("duration"), out oDuration);
-                    this.acodec = node.GetAttribute("acodec");
-                    Int32.TryParse(node.GetAttribute("asampfreq"), out oAsampfreq);
-                    this.asampfmt = node.GetAttribute("asampfmt");
-                    Int32.TryParse(node.GetAttribute("abitrate"), out oAbitrate);
-                    this.size = oSize;
-                    this.width = oWidth;
-                    this.height = oHeight;
-                    this.duration = oDuration;
-                    this.asampfreq = oAsampfreq;
-                    this.abitrate = oAbitrate;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Retrieve messages from the server
-        /// </summary>
         public void PollMessages(bool autoReceipt = true)
         {
             while (pollMessage(autoReceipt)) ;
@@ -727,10 +382,11 @@ namespace WhatsAppApi
             return false;
         }
 
-        /// <summary>
-        /// Add the authenication nodes
-        /// </summary>
-        /// <returns>An instance of the ProtocolTreeNode class</returns>
+        protected ProtocolTreeNode addFeatures()
+        {
+            return new ProtocolTreeNode("stream:features", null);
+        }
+
         protected ProtocolTreeNode addAuth()
         {
             List<KeyValue> attr = new List<KeyValue>(new KeyValue[] {
@@ -776,10 +432,6 @@ namespace WhatsAppApi
             return data;
         }
 
-        /// <summary>
-        /// Add the auth response to protocoltreenode
-        /// </summary>
-        /// <returns>An instance of the ProtocolTreeNode</returns>
         protected ProtocolTreeNode addAuthResponse()
         {
             if (this._challengeBytes != null)
@@ -806,40 +458,11 @@ namespace WhatsAppApi
             throw new Exception("Auth response error");
         }
 
-        /// <summary>
-        /// Add stream features
-        /// </summary>
-        /// <returns></returns>
-        protected ProtocolTreeNode addFeatures()
-        {
-            return new ProtocolTreeNode("stream:features", null);
-        }
-
-        /// <summary>
-        /// Print a message to the debug console
-        /// </summary>
-        /// <param name="debugMsg">The message</param>
-        protected void DebugPrint(string debugMsg)
-        {
-            if (WhatsApp.DEBUG && debugMsg.Length > 0)
-            {
-                Console.WriteLine(debugMsg);
-            }
-        }
-
-        /// <summary>
-        /// Process the challenge
-        /// </summary>
-        /// <param name="node">The node that contains the challenge</param>
         protected void processChallenge(ProtocolTreeNode node)
         {
             _challengeBytes = node.data;
         }
         
-        /// <summary>
-        /// Process inbound data
-        /// </summary>
-        /// <param name="data">Data to process</param>
         protected bool processInboundData(byte[] msgdata, bool autoReceipt = true)
         {
             try
@@ -858,23 +481,13 @@ namespace WhatsAppApi
                                                            node.GetAttribute("kind"),
                                                            node.GetAttribute("creation"),
                                                            node.GetAttribute("expiration"));
-                        if (this.OnLoginSuccess != null)
-                        {
-                            this.OnLoginSuccess(this.phoneNumber, node.GetData());
-                        }
+                        this.fireOnLoginSuccess(this.phoneNumber, node.GetData());
                     }
                     else if (ProtocolTreeNode.TagEquals(node, "failure"))
                     {
                         this.loginStatus = CONNECTION_STATUS.UNAUTHORIZED;
-                        if (this.OnLoginFailed != null)
-                        {
-                            this.OnLoginFailed(node.children.FirstOrDefault().tag);
-                        }
+                        this.fireOnLoginFailed(node.children.FirstOrDefault().tag);
                     }
-                    
-                    
-                    
-                    
                     
                     if (ProtocolTreeNode.TagEquals(node, "receipt"))
                     {
@@ -885,10 +498,7 @@ namespace WhatsAppApi
                         {
                             case "delivery":
                                 //delivered to target
-                                if (this.OnGetMessageReceivedClient != null)
-                                {
-                                    this.OnGetMessageReceivedClient(from, id);
-                                }
+                                this.fireOnGetMessageReceivedClient(from, id);
                                 break;
                             case "read":
                                 //read by target
@@ -904,7 +514,6 @@ namespace WhatsAppApi
                         SendNotificationAck(node, type);
                     }
 
-
                     if (ProtocolTreeNode.TagEquals(node, "message"))
                     {
                         this.handleMessage(node, autoReceipt);
@@ -915,8 +524,6 @@ namespace WhatsAppApi
                     {
                         this.handleIq(node);
                     }
-
-
 
                     if (ProtocolTreeNode.TagEquals(node, "stream:error"))
                     {
@@ -929,18 +536,11 @@ namespace WhatsAppApi
                         this.Disconnect();
                     }
 
-
-
                     if (ProtocolTreeNode.TagEquals(node, "presence"))
                     {
                         //presence node
-                        if (this.OnGetPresence != null)
-                        {
-                            this.OnGetPresence(node.GetAttribute("from"), node.GetAttribute("type"));
-                        }
+                        this.fireOnGetPresence(node.GetAttribute("from"), node.GetAttribute("type"));
                     }
-
-
 
                     if (node.tag == "ib")
                     {
@@ -960,31 +560,21 @@ namespace WhatsAppApi
                         }
                     }
 
-
-
                     if (node.tag == "chatstate")
                     {
                         string state = node.children.FirstOrDefault().tag;
                         switch (state)
                         {
                             case "composing":
-                                if (this.OnGetTyping != null)
-                                {
-                                    this.OnGetTyping(node.GetAttribute("from"));
-                                }
+                                this.fireOnGetTyping(node.GetAttribute("from"));
                                 break;
                             case "paused":
-                                if (this.OnGetPaused != null)
-                                {
-                                    this.OnGetPaused(node.GetAttribute("from"));
-                                }
+                                this.fireOnGetPaused(node.GetAttribute("from"));
                                 break;
                             default:
                                 throw new NotImplementedException(node.NodeString());
                         }
                     }
-
-
 
                     if (node.tag == "ack")
                     {
@@ -992,26 +582,14 @@ namespace WhatsAppApi
                         if (cls == "message")
                         {
                             //server receipt
-                            if (this.OnGetMessageReceivedServer != null)
-                            {
-                                this.OnGetMessageReceivedServer(node.GetAttribute("from"), node.GetAttribute("id"));
-                            }
-                        }
-                        else if(cls == "receipt")
-                        {
-                            //message received ack
-                            //do nothing
+                            this.fireOnGetMessageReceivedServer(node.GetAttribute("from"), node.GetAttribute("id"));
                         }
                     }
-
-
 
                     if (node.tag == "notification")
                     {
                         this.handleNotification(node);
                     }
-
-
 
                     return true;
                 }
@@ -1028,10 +606,7 @@ namespace WhatsAppApi
             if (!string.IsNullOrEmpty(node.GetAttribute("notify")))
             {
                 string name = node.GetAttribute("notify");
-                if (this.OnGetContactName != null)
-                {
-                    this.OnGetContactName(node.GetAttribute("from"), name);
-                }
+                this.fireOnGetContactName(node.GetAttribute("from"), name);
             }
             if (node.GetAttribute("type") == "error")
             {
@@ -1040,10 +615,7 @@ namespace WhatsAppApi
             if (node.GetChild("body") != null)
             {
                 //text message
-                if (this.OnGetMessage != null)
-                {
-                    this.OnGetMessage(node, node.GetAttribute("from"), node.GetAttribute("id"), node.GetAttribute("notify"), System.Text.Encoding.UTF8.GetString(node.GetChild("body").GetData()), autoReceipt);
-                }
+                this.fireOnGetMessage(node, node.GetAttribute("from"), node.GetAttribute("id"), node.GetAttribute("notify"), System.Text.Encoding.UTF8.GetString(node.GetChild("body").GetData()), autoReceipt);
                 if (autoReceipt)
                 {
                     this.sendMessageReceived(node);
@@ -1063,54 +635,39 @@ namespace WhatsAppApi
                 switch (media.GetAttribute("type"))
                 {
                     case "image":
-                        if (this.OnGetMessageImage != null)
-                        {
-                            url = media.GetAttribute("url");
-                            file = media.GetAttribute("file");
-                            size = Int32.Parse(media.GetAttribute("size"));
-                            preview = media.GetData();
-                            this.OnGetMessageImage(from, id, file, size, url, preview);
-                        }
+                        url = media.GetAttribute("url");
+                        file = media.GetAttribute("file");
+                        size = Int32.Parse(media.GetAttribute("size"));
+                        preview = media.GetData();
+                        this.fireOnGetMessageImage(from, id, file, size, url, preview);
                         break;
                     case "audio":
-                        if (this.OnGetMessageAudio != null)
-                        {
-                            file = media.GetAttribute("file");
-                            size = Int32.Parse(media.GetAttribute("size"));
-                            url = media.GetAttribute("url");
-                            preview = media.GetData();
-                            this.OnGetMessageAudio(from, id, file, size, url, preview);
-                        }
+                        file = media.GetAttribute("file");
+                        size = Int32.Parse(media.GetAttribute("size"));
+                        url = media.GetAttribute("url");
+                        preview = media.GetData();
+                        this.fireOnGetMessageAudio(from, id, file, size, url, preview);
                         break;
                     case "video":
-                        if (this.OnGetMessageVideo != null)
-                        {
-                            file = media.GetAttribute("file");
-                            size = Int32.Parse(media.GetAttribute("size"));
-                            url = media.GetAttribute("url");
-                            preview = media.GetData();
-                            this.OnGetMessageVideo(from, id, file, size, url, preview);
-                        }
+                        file = media.GetAttribute("file");
+                        size = Int32.Parse(media.GetAttribute("size"));
+                        url = media.GetAttribute("url");
+                        preview = media.GetData();
+                        this.fireOnGetMessageVideo(from, id, file, size, url, preview);
                         break;
                     case "location":
-                        if (this.OnGetMessageLocation != null)
-                        {
-                            double lon = double.Parse(media.GetAttribute("longitude"), System.Globalization.CultureInfo.InvariantCulture);
-                            double lat = double.Parse(media.GetAttribute("latitude"), System.Globalization.CultureInfo.InvariantCulture);
-                            preview = media.GetData();
-                            name = media.GetAttribute("name");
-                            url = media.GetAttribute("url");
-                            this.OnGetMessageLocation(from, id, lon, lat, url, name, preview);
-                        }
+                        double lon = double.Parse(media.GetAttribute("longitude"), System.Globalization.CultureInfo.InvariantCulture);
+                        double lat = double.Parse(media.GetAttribute("latitude"), System.Globalization.CultureInfo.InvariantCulture);
+                        preview = media.GetData();
+                        name = media.GetAttribute("name");
+                        url = media.GetAttribute("url");
+                        this.fireOnGetMessageLocation(from, id, lon, lat, url, name, preview);
                         break;
                     case "vcard":
-                        if (this.OnGetMessageVcard != null)
-                        {
-                            ProtocolTreeNode vcard = media.GetChild("vcard");
-                            name = vcard.GetAttribute("name");
-                            dat = vcard.GetData();
-                            this.OnGetMessageVcard(from, id, name, dat);
-                        }
+                        ProtocolTreeNode vcard = media.GetChild("vcard");
+                        name = vcard.GetAttribute("name");
+                        dat = vcard.GetData();
+                        this.fireOnGetMessageVcard(from, id, name, dat);
                         break;
                 }
                 this.sendMessageReceived(node);
@@ -1121,10 +678,7 @@ namespace WhatsAppApi
         {
             if (node.GetAttribute("type") == "error")
             {
-                if (this.OnError != null)
-                {
-                    this.OnError(node.GetAttribute("id"), node.GetAttribute("from"), Int32.Parse(node.GetChild("error").GetAttribute("code")), node.GetChild("error").GetAttribute("text"));
-                }
+                this.fireOnError(node.GetAttribute("id"), node.GetAttribute("from"), Int32.Parse(node.GetChild("error").GetAttribute("code")), node.GetChild("error").GetAttribute("text"));
             }
             if (node.GetChild("sync") != null)
             {
@@ -1146,10 +700,7 @@ namespace WhatsAppApi
                 }
                 int index = 0;
                 Int32.TryParse(sync.GetAttribute("index"), out index);
-                if (this.OnGetSyncResult != null)
-                {
-                    this.OnGetSyncResult(index, sync.GetAttribute("sid"), existingUsers, failedNumbers.ToArray());
-                }
+                this.fireOnGetSyncResult(index, sync.GetAttribute("sid"), existingUsers, failedNumbers.ToArray());
             }
             if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
                 && node.children.FirstOrDefault().tag == "query"
@@ -1158,10 +709,7 @@ namespace WhatsAppApi
             {
                 //last seen
                 DateTime lastSeen = DateTime.Now.AddSeconds(double.Parse(node.children.FirstOrDefault().GetAttribute("seconds")) * -1);
-                if (this.OnGetLastSeen != null)
-                {
-                    this.OnGetLastSeen(node.GetAttribute("from"), lastSeen);
-                }
+                this.fireOnGetLastSeen(node.GetAttribute("from"), lastSeen);
             }
             if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
                 && (ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "media") || ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "duplicate"))
@@ -1181,17 +729,11 @@ namespace WhatsAppApi
                 string type = node.GetChild("picture").GetAttribute("type");
                 if (type == "preview")
                 {
-                    if (this.OnGetPhotoPreview != null)
-                    {
-                        this.OnGetPhotoPreview(from, id, dat);
-                    }
+                    this.fireOnGetPhotoPreview(from, id, dat);
                 }
                 else
                 {
-                    if (this.OnGetPhoto != null)
-                    {
-                        this.OnGetPhoto(from, id, dat);
-                    }
+                    this.fireOnGetPhoto(from, id, dat);
                 }
             }
             if (node.GetAttribute("type").Equals("get", StringComparison.OrdinalIgnoreCase)
@@ -1203,10 +745,10 @@ namespace WhatsAppApi
                 && ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "group"))
             {
                 //group(s) info
-                List<GroupInfo> groups = new List<GroupInfo>();
+                List<WaGroupInfo> groups = new List<WaGroupInfo>();
                 foreach (ProtocolTreeNode group in node.children)
                 {
-                    groups.Add(new GroupInfo(
+                    groups.Add(new WaGroupInfo(
                         group.GetAttribute("id"),
                         group.GetAttribute("owner"),
                         long.Parse(group.GetAttribute("creation")),
@@ -1215,10 +757,7 @@ namespace WhatsAppApi
                         group.GetAttribute("s_o")
                         ));
                 }
-                if (this.OnGetGroups != null)
-                {
-                    this.OnGetGroups(groups.ToArray());
-                }
+                this.fireOnGetGroups(groups.ToArray());
             }
             if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
                 && ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "participant"))
@@ -1232,10 +771,7 @@ namespace WhatsAppApi
                         participants.Add(part.GetAttribute("jid"));
                     }
                 }
-                if (this.OnGetGroupParticipants != null)
-                {
-                    this.OnGetGroupParticipants(node.GetAttribute("from"), participants.ToArray());
-                }
+                this.fireOnGetGroupParticipants(node.GetAttribute("from"), participants.ToArray());
             }
         }
 
@@ -1243,27 +779,18 @@ namespace WhatsAppApi
         {
             if (!String.IsNullOrEmpty(node.GetAttribute("notify")))
             {
-                if (this.OnGetContactName != null)
-                {
-                    this.OnGetContactName(node.GetAttribute("from"), node.GetAttribute("notify"));
-                }
+                this.fireOnGetContactName(node.GetAttribute("from"), node.GetAttribute("notify"));
             }
             string type = node.GetAttribute("type");
             switch (type)
             {
                 case "picture":
                     ProtocolTreeNode child = node.children.FirstOrDefault();
-                    if (this.OnNotificationPicture != null)
-                    {
-                        this.OnNotificationPicture(child.tag, child.GetAttribute("jid"), child.GetAttribute("id"));
-                    }
+                    this.fireOnNotificationPicture(child.tag, child.GetAttribute("jid"), child.GetAttribute("id"));
                     break;
                 case "status":
                     ProtocolTreeNode child2 = node.children.FirstOrDefault();
-                    if (this.OnGetStatus != null)
-                    {
-                        this.OnGetStatus(node.GetAttribute("from"), child2.tag, node.GetAttribute("notify"), System.Text.Encoding.UTF8.GetString(child2.GetData()));
-                    }
+                    this.fireOnGetStatus(node.GetAttribute("from"), child2.tag, node.GetAttribute("notify"), System.Text.Encoding.UTF8.GetString(child2.GetData()));
                     break;
                 default:
                     throw new NotImplementedException(node.NodeString());
@@ -1350,87 +877,6 @@ namespace WhatsAppApi
         {
             this.DebugPrint(p);
         }
-
-
-
-
-
-        // events
-        public event NullDelegate OnConnectSuccess;
-        public event ExceptionDelegate OnConnectFailed;
-        public event ExceptionDelegate OnDisconnect;
-        public event LoginSuccessDelegate OnLoginSuccess;
-        public event StringDelegate OnLoginFailed;
-
-        public event OnGetMessageDelegate OnGetMessage;
-        public event OnGetMediaDelegate OnGetMessageImage;
-        public event OnGetMediaDelegate OnGetMessageVideo;
-        public event OnGetMediaDelegate OnGetMessageAudio;
-        public event OnGetLocationDelegate OnGetMessageLocation;
-        public event OnGetVcardDelegate OnGetMessageVcard;
-
-        public event OnErrorDelegate OnError;
-        public event OnNotificationPictureDelegate OnNotificationPicture;
-        
-        public event OnGetMessageReceivedDelegate OnGetMessageReceivedServer;
-        public event OnGetMessageReceivedDelegate OnGetMessageReceivedClient;
-
-        public event OnGetPresenceDelegate OnGetPresence;
-        public event OnGetGroupParticipantsDelegate OnGetGroupParticipants;
-        public event OnGetLastSeenDelegate OnGetLastSeen;
-        public event OnGetchatStateDelegate OnGetTyping;
-        public event OnGetchatStateDelegate OnGetPaused;
-        public event OnGetPictureDelegate OnGetPhoto;
-        public event OnGetPictureDelegate OnGetPhotoPreview;
-        public event OnGetGroupsDelegate OnGetGroups;
-        public event OnContactNameDelegate OnGetContactName;
-        public event OnGetStatusDelegate OnGetStatus;
-
-        public event OnGetSyncResultDelegate OnGetSyncResult;
-
-        //event delegates
-        public delegate void OnContactNameDelegate(string from, string contactName);
-        public delegate void NullDelegate();
-        public delegate void ExceptionDelegate(Exception ex);
-        public delegate void LoginSuccessDelegate(string phoneNumber, byte[] data);
-        public delegate void StringDelegate(string data);
-        public delegate void OnErrorDelegate(string id, string from, int code, string text);
-        public delegate void OnGetMessageReceivedDelegate(string from, string id);
-        public delegate void OnNotificationPictureDelegate(string type, string jid, string id);
-        public delegate void OnGetMessageDelegate(ProtocolTreeNode messageNode, string from, string id, string name, string message, bool receipt_sent);
-        public delegate void OnGetPresenceDelegate(string from, string type);
-        public delegate void OnGetGroupParticipantsDelegate(string gjid, string[] jids);
-        public delegate void OnGetLastSeenDelegate(string from, DateTime lastSeen);
-        public delegate void OnGetchatStateDelegate(string from);
-        public delegate void OnGetMediaDelegate(string from, string id, string fileName, int fileSize, string url, byte[] preview);
-        public delegate void OnGetLocationDelegate(string from, string id, double lon, double lat, string url, string name, byte[] preview);
-        public delegate void OnGetVcardDelegate(string from, string id, string name, byte[] data);
-        public delegate void OnGetPictureDelegate(string from, string id, byte[] data);
-        public delegate void OnGetGroupsDelegate(GroupInfo[] groups);
-        public delegate void OnGetStatusDelegate(string from, string type, string name, string status);
-        public delegate void OnGetSyncResultDelegate(int index, string sid, Dictionary<string, string> existingUsers, string[] failedNumbers);
-
-        public class GroupInfo
-        {
-            public readonly string id;
-            public readonly string owner;
-            public readonly long creation;
-            public readonly string subject;
-            public readonly long subjectChangedTime;
-            public readonly string subjectChangedBy;
-
-            internal GroupInfo(string id, string owner, long creation, string subject, long subjectChanged, string subjectChangedBy)
-            {
-                this.id = id;
-                this.owner = owner;
-                this.creation = creation;
-                this.subject = subject;
-                this.subjectChangedTime = subjectChanged;
-                this.subjectChangedBy = subjectChangedBy;
-            }
-        }
-
-        internal BinTreeNodeWriter BinWriter;
 
         public void SendActive()
         {
