+++ WhatsAppApi/WhatsAppApi.csproj
@@ -75,10 +75,14 @@
     <Compile Include="Register\WaToken.cs" />
     <Compile Include="Register\WhatsRegisterV2.cs" />
     <Compile Include="Response\CorruptStreamException.cs" />
+    <Compile Include="Response\WaGroupInfo.cs" />
+    <Compile Include="Response\WaUploadResponse.cs" />
     <Compile Include="Response\WhatsEventHandler.cs" />
     <Compile Include="Settings\WhatsConstants.cs" />
     <Compile Include="WhatsApp.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="WhatsAppBase.cs" />
+    <Compile Include="WhatsEventBase.cs" />
     <Compile Include="WhatsNetwork.cs" />
   </ItemGroup>
   <ItemGroup>
